Replaced legacy 'float' with CGFloat where appropriate.

+++ cocos2d/CCActionEase.m
@@ -36,7 +36,7 @@
 #import "CCActionEase.h"
 
 #ifndef M_PI_X_2
-#define M_PI_X_2 (float)M_PI * 2.0
+#define M_PI_X_2 M_PI * 2.0
 #endif
 
 #pragma mark EaseAction
@@ -98,7 +98,7 @@
 @implementation CCActionEaseSineIn
 -(void) update: (CCTime) t
 {
-	[_inner update:-1*cosf(t * (float)M_PI_2) +1];
+	[_inner update:-1*cosf(t * M_PI_2) +1];
 }
 
 - (CCActionInterval*) reverse
@@ -113,7 +113,7 @@
 @implementation CCActionEaseSineOut
 -(void) update: (CCTime) t
 {
-	[_inner update:sinf(t * (float)M_PI_2)];
+	[_inner update:sinf(t * M_PI_2)];
 }
 
 - (CCActionInterval*) reverse
@@ -128,7 +128,7 @@
 @implementation CCActionEaseSineInOut
 -(void) update: (CCTime) t
 {
-	[_inner update:-0.5*(cosf( (float)M_PI*t) - 1)];
+	[_inner update:-0.5*(cosf( M_PI*t) - 1)];
 }
 @end
 
@@ -138,12 +138,12 @@
 
 @implementation CCActionEaseRate
 @synthesize rate=_rate;
-+(instancetype) actionWithAction: (CCActionInterval*) action rate:(float)rate
++(instancetype) actionWithAction: (CCActionInterval*) action rate:(CGFloat)rate
 {
 	return [[self alloc] initWithAction: action rate:rate];
 }
 
--(id) initWithAction: (CCActionInterval*) action rate:(float)rate
+-(id) initWithAction: (CCActionInterval*) action rate:(CGFloat)rate
 {
 	if( (self=[super initWithAction:action ]) )
 		self.rate = rate;
@@ -214,7 +214,7 @@
 	return [[self alloc] initWithAction:action period:0.3];
 }
 
-+(instancetype) actionWithAction: (CCActionInterval*) action period:(float)period
++(instancetype) actionWithAction: (CCActionInterval*) action period:(CGFloat)period
 {
 	return [[self alloc] initWithAction:action period:period];
 }
@@ -224,7 +224,7 @@
 	return [self initWithAction:action period:0.3];
 }
 
--(id) initWithAction: (CCActionInterval*) action period:(float)period
+-(id) initWithAction: (CCActionInterval*) action period:(CGFloat)period
 {
 	if( (self=[super initWithAction:action]) )
 		_period = period;
@@ -255,7 +255,7 @@
 		newT = t;
 
 	else {
-		float s = _period / 4;
+		CGFloat s = _period / 4;
 		t = t - 1;
 		newT = -powf(2, 10 * t) * sinf( (t-s) *M_PI_X_2 / _period);
 	}
@@ -279,7 +279,7 @@
 		newT = t;
 
 	} else {
-		float s = _period / 4;
+		CGFloat s = _period / 4;
 		newT = powf(2, -10 * t) * sinf( (t-s) *M_PI_X_2 / _period) + 1;
 	}
 	[_inner update:newT];
